third_party/blink/web_tests/wpt_internal/fenced_frame/opaque-ad-sizes.https.html

<!DOCTYPE html>
<title>Test frame size restrictions in opaque ads mode.</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="/common/dispatcher/dispatcher.js"></script>
<script src="resources/utils.js"></script>
<meta name="timeout" content="long">

<body>
<script>
async function runTest(input_width, input_height, output_width, output_height) {
  // Attach a fenced frame whose outer container has dimensions
  // `input_width` by `input_height`.
  const frame = attachFencedFrameContext({attributes: [["mode", "opaque-ads"],
      ["width", input_width], ["height", input_height]]});

  const assert_dimensions =
      (label, input_width, input_height, output_width, output_height) => {
    assert_equals(window.innerWidth, output_width,
        label + " the innerWidth " + input_width + " coerces to " + output_width);
    assert_equals(getComputedStyle(document.documentElement).width,
        output_width+"px",
        label + " the computed width coerces to " + output_width);
    assert_equals(window.innerHeight, output_height,
        label + " the innerHeight " + input_height + " coerces to " + output_height);
  }

  // Assert that the fenced frame sees its dimensions rounded to the nearest
  // ad size.
  await frame.execute(assert_dimensions,
      ["After navigation", input_width, input_height, output_width, output_height]);

  // Assert that the embedder sees the fenced frame's original dimensions.
  assert_equals(frame.width, input_width.toString(),
      "The outer container width is the requested width.");
  assert_equals(frame.height, input_height.toString(),
      "The outer container height is the requested height.");

  // Resize the fenced frame's outer container.
  const new_size_x = 320;
  const new_size_y = 50;
  frame.width = new_size_x;
  frame.height = new_size_y;

  // Refresh the fenced frame.
  await frame.execute(() => {
    window.executor.suspend(() => {
      location.href = location.href;
    });
  });

  // Observe that navigations after the first don't change the fenced frame's
  // inner dimensions.
  await frame.execute(assert_dimensions,
      ["After resizing", input_width, input_height, output_width, output_height]);
}

// Exact size cases.
promise_test(async () => { return runTest(320, 50, 320, 50); }, '320x50');
promise_test(async () => { return runTest(728, 90, 728, 90); }, '728x90');
promise_test(async () => { return runTest(970, 90, 970, 90); }, '970x90');
promise_test(async () => { return runTest(320, 100, 320, 100); }, '320x100');
promise_test(async () => { return runTest(160, 600, 160, 600); }, '160x600');
promise_test(async () => { return runTest(300, 250, 300, 250); }, '300x250');
promise_test(async () => { return runTest(970, 250, 970, 250); }, '970x250');
promise_test(async () => { return runTest(336, 280, 336, 280); }, '336x280');
promise_test(async () => { return runTest(320, 480, 320, 480); }, '320x48');
promise_test(async () => { return runTest(300, 600, 300, 600); }, '300x600');
promise_test(async () => { return runTest(300, 1050, 300, 1050); }, '300x1050');

// Rounding cases.
promise_test(async () => { return runTest(970.1, 250, 970, 250); }, '970.1x250');
promise_test(async () => { return runTest(970, 250.1, 970, 250); }, '970x250.1');
promise_test(async () => { return runTest(971, 250, 970, 250); }, '971x250');
promise_test(async () => { return runTest(970, 251, 970, 250); }, '970x251');

// Edge cases. The particular sizes it rounds to aren't important here, just
// that it rounds to one of the sizes on the allowed list.
promise_test(async () => { return runTest(0, 100, 320, 50); }, '0x100');
promise_test(async () => { return runTest(100, 0, 320, 50); }, '100x0');
promise_test(async () => { return runTest(Number.MAX_VALUE, Number.MAX_VALUE, 300, 250); }, 'MAXxMAX');
promise_test(async () => { return runTest(Number.POSITIVE_INFINITY, Number.POSITIVE_INFINITY, 320, 100); }, 'INFINITYxINFINITY');
</script>
</body>
